requirements.txt
alphabet-detector==0.0.7
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
black==22.6.0
certifi==2022.5.18.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
drf-yasg==1.21.3
Faker==13.15.1
filelock==3.7.1
gunicorn==20.1.0
identify==2.5.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
model-bakery==1.7.0
msal==1.18.0
mypy-extensions==0.4.3
nodeenv==1.6.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
